+++ package.json
@@ -37,6 +37,7 @@
     "whatwg-fetch": "^3.6.2"
   },
   "dependencies": {
+    "native-html-tabs": "^0.1.4",
     "regenerator-runtime": "^0.13.11"
   }
 }
