+++ src/templates/index.html
@@ -26,8 +26,8 @@
     </nav>
   </header>
 
-  <div id="content">
-  </div>
+  <main id="content">
+  </main>
 </body>
 
 </html>
